// src/modules/inventory/reports/yarn-store/yarn-rcv-issue-register-report/yarn-rcv-issue-register-report-index-type.ts
export interface YarnRcvIssueRegisterReportType {
  LC_ID: number;
  WORK_ORDER_ID: number;
  LC_NO: string;
  YARN_ID: number;
  YARN: string;
  BRAND: string;
  YARN_RECEIVED_DATE: string; // ISO date string
  CHALLAN_NO: string;
  YARN_LOT_NUMBER_ID: number;
  YARN_LOT_NUMBER: string;
  KNITTING_HOUSE_ID: number;
  KNITTING_HOUSE: string;
  IMPORT_TYPE: string;
  INTER_COMPANY_RCV: string;
  RECEIVE_CTN: number;
  RECEIVE_KG: number;
  PER_CTN_QTY: string;
  INTER_COMPANY_ISSUE: string;
  ISSUE_CTN: number;
  ISSUE_KG: number;
  PO_NO: string;
  SERIAL_OF_ROW: number;
  DATA_SOURCE: string;
  BUYER: string;
  TOTAL_BALANCE: number;
  TOTAL_BALANCE_CTN: number;
  ID: number;
  COMPANY_NAME: string;
  COMPANY_ADDRESS: string;
  COMPANY_CONTACT: string;
  COMPANY_EMAIL: string;
  COMPANY_WEB: string;
  COMPANY_REMARKS: string;
  CHECK_COLUMN: string;
  BALANCE_QUANTITY: number;
  YARN_2: string;
}